"Enter" toggles a Luminance histogram
There is a strange bug. When at the luminance histogram
and the video pauses, the histogram blinks on/off...
This does not happen with the RGB histograms :-/

After some experimantation, it seems that:
When on pause, the current frame is run through the filter
chain continuosly in a loop. Generally, filters paint
on a new picture and do not touch their current(input) frame.
But this one does, so it keeps painting on the same area,
each time reversing the colors...